Refactored: RegisterForm moved to separated file, barrel Form exports.
